backend/requirements.txt
vnstock
fastapi
uvicorn[standard]
fastapi[standard]
numpy
pandas
ta-lib
pydantic
sqlalchemy
psycopg2-binary
uvicorn
passlib
bcrypt
asyncio
PyJWT
httpx